# Improve pouchDB performances