We noticed severe performances issues when using pouchdb in a  react-native environment.
See commit descriptions for more insights

**BEFORE**
```
Indexing time : 53 868 ms
Query time : 51 499 ms

```
**AFTER**
```
Indexing time : 31 833 ms
Query time : 1 854 ms