@paultranvan 2024-10-04
> 30s to index 40k documents is still huge :(

It is indeed. While this time would be dramatically reduced on a more powerful device, this is disappointing, but not so surprising regarding previous performances measures, that tend to show massive performances differences depending on the adapter and document size : 
![image](https://github.com/user-attachments/assets/d0f47ae3-aa40-4c1b-beb0-d47da1ce01e4)